requirements.txt
# Core ML
scikit-learn==1.3.0
pandas==1.5.3
numpy==1.24.3
joblib==1.3.2

# API
fastapi==0.99.1
uvicorn==0.24.0

# Streaming
kafka-python==2.0.2

# Model Registry
mlflow==2.9.2

# Monitoring
evidently==0.5.1
streamlit==1.31.0
plotly>=5.18.0

# Explainability
shap==0.43.0
lime==0.2.0.1

# Utils
python-dotenv==1.0.0
requests>=2.31.0